categories_api/requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
Django==2.2.7
djangorestframework==3.10.3
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
psycopg2-binary==2.8.4
py==1.8.0
pyparsing==2.4.5
pytest==5.2.4
pytest-django==3.7.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
wcwidth==0.1.7
zipp==0.6.0
